requirements.txt
aiogram==3.22.0
aiohttp==3.12.15
asyncpg==0.30.0
greenlet>=3.0.0
loguru==0.7.3
pydantic==2.11.10
pydantic-settings==2.11.0
python-dotenv==1.1.1
PyYAML==6.0.3
remnawave==2.3.2
SQLAlchemy==2.0.44
